BREAKING CHANGE: change structure proyect and now download files
