+++ static/css/style.css
@@ -1,0 +1,322 @@
+#scrollToTopBtn {
+    display: none; /* Ocultar el botón por defecto */
+    position: fixed;
+    bottom: 20px;
+    right: 30px;
+    z-index: 99;
+    font-size: 18px;
+    border: none;
+    outline: none;
+    background-color: #555;
+    color: white;
+    cursor: pointer;
+    padding: 15px;
+    border-radius: 4px;
+}
+
+.tab {
+    overflow: hidden;
+    display: flex;
+    justify-content: center; /* Centra horizontalmente */
+    border: 1px solid #ccc;
+    background-color: #f1f1f1;
+}
+
+.tab button {
+    background-color: inherit;
+    float: left;
+    border: none;
+    outline: none;
+    cursor: pointer;
+    padding: 14px 16px;
+    transition: 0.3s;
+}
+
+.tab button:hover {
+    background-color: #ddd;
+}
+
+.tab button.active {
+    background-color: #ccc;
+}
+
+.tabcontent {
+    display: none;
+    padding: 6px 12px;
+    border: 1px solid #ccc;
+    border-top: none;
+}
+
+.select-button-container {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    padding: 20px;
+    border: 1px solid #ccc;
+    border-radius: 10px;
+    background-color: #f9f9f9;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    gap: 10px; /* Espacio entre los elementos */
+}
+
+.select-button-container label {
+    margin-right: 10px; /* Espacio entre el label y el primer select */
+}
+
+/*Select para ver tablas*/
+.table-select { 
+    width: 30%;
+    padding: 5px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    background-color: #e9e9e9;
+    font-size: 14px;
+    margin-right: 10px; /* Espacio entre los selects */
+}
+
+
+.extra-button {
+    margin-left: 10px;
+    padding: 10px 20px;
+    background-color: #fe001a; /* Rojo cocacola por defecto */
+    color: white;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+}
+
+.extra-button:hover {
+    background-color: #c70000;
+}
+
+/* General table styling */
+.editable-table {
+    width:  100%;
+    border-collapse: collapse;
+    font-family: Arial, sans-serif;
+    margin: 15px 0;
+    font-size: 14px;
+    text-align: center;
+    background-color: #f9f9f9;
+    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.15);
+    table-layout: auto; /* Ajusta automáticamente el ancho de las columnas */
+}
+
+/* Estilo para los encabezados */
+.editable-table-header th {
+    background-color: #d6d6d6;
+    font-size: 18px;
+    color: rgb(63, 63, 63);
+    padding: 8px;
+    position: sticky;
+    top: 0;
+}
+
+/* Estilo para el cuerpo de la tabla */
+.editable-table-body td {
+    padding: 6px;
+    border: 1px solid #ddd;
+    vertical-align: middle;
+}
+
+/* Estilo para inputs de celdas */
+.editable-table-body input[type="text"] {
+    width: auto;
+    padding: 6px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+}
+
+/* Estilo para selects */
+.editable-table-body select {
+    width: auto;
+    padding: 6px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    background-color: #fff;
+    font-size: 14px;
+}
+
+
+.checkbox-cell {
+    max-width: 10px;
+    padding: 5px;
+    text-align: center;
+    align-items: center;
+    
+}
+
+/* Estilo para el botón del filtro */
+.filter-container .dropdown-button {
+    background-color: #fe001a;
+    color: white;
+    border: none;
+    padding: 2px 2px;
+    border-radius: 4px;
+    cursor: pointer;
+    font-size: 14px;
+}
+
+.filter-container .dropdown-button:hover {
+    background-color: #c70000;
+}
+
+/* Estilo para el menú desplegable */
+.filter-container .dropdown-menu {
+    display: none;
+    position: absolute;
+    background-color: #fff;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    z-index: 1;
+    margin-top: 5px;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+    padding: 10px;
+    max-height: 400px;
+    overflow-y: auto;
+
+}
+
+.filter-container .dropdown-menu.show {
+    display: block;
+}
+
+/* Estilo para las opciones del menú */
+.filter-container .dropdown-item {
+    display: flex;
+    align-items: center;
+    padding: 5px 10px;
+    cursor: pointer;
+}
+
+.filter-container .dropdown-item input[type="checkbox"] {
+    margin-right: 6px;
+}
+
+/* Botón para borrar filtros */
+.clear-filters-button {
+    display: block;
+    background-color: #f44336;
+    color: white;
+    border: none;
+    padding: 5px 10px;
+    border-radius: 4px;
+    cursor: pointer;
+    font-size: 12px;
+    margin-bottom: 10px;
+}
+
+.clear-filters-button:hover {
+    background-color: #e53935;
+}
+
+
+/* Alternar color de filas */
+.editable-table-body tr:nth-child(even) {
+    background-color: #f9f9f9;
+}
+
+.editable-table-body tr:nth-child(odd) {
+    background-color: #fff;
+}
+
+.clear-filters-button {
+    width: 100%;
+    margin-bottom: 5px;
+    padding: 5px;
+    border-radius: 5px;
+    background-color: #fe001a;
+    color: white;
+    border: 1px solid #ccc;
+    cursor: pointer;
+}
+
+.clear-filters-button:hover {
+    background-color: #c70000;
+}
+
+
+.delete-column-button {
+    margin-bottom: 10px;
+    padding: 5px 10px;
+    background-color: #f44336;
+    color: white;
+    border: none;
+    border-radius: 3px;
+    cursor: pointer;
+}
+
+.delete-column-button:hover {
+    background-color: #d32f2f;
+}
+
+
+.hidden-column {
+    display: none;
+}
+
+
+.delete-button {
+    padding: 5px 10px;
+    font-size: 14px;
+    background-color: #ff6666;
+    color: white;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    
+}
+
+.delete-button:hover {
+    background-color: #ff3333;
+}
+
+
+/* Contenedor del select */
+.filter-select {
+    width: 100%; /* Ancho completo */
+    height: 85px; /* Alto del select para ver varias opciones */
+    padding: 6px; /* Espaciado interno */
+    border: 2px solid #ccc; /* Borde del select */
+    border-radius: 5px; /* Bordes redondeados */
+    background-color: #f9f9f9; /* Color de fondo */
+    font-size: 14px; /* Tamaño de la fuente */
+    color: #333; /* Color del texto */
+    font-family: Arial, sans-serif; /* Tipo de letra */
+    box-sizing: border-box; /* Asegura que el padding no afecte el tamaño */
+    
+}
+
+
+/* Cambiar el fondo de la opción seleccionada */
+.filter-select option:checked {
+    background-color: #ccc; /* Fondo personalizado */
+    color: rgb(46, 46, 46); /* Texto */
+}
+
+
+/* Estilo cuando el select está enfocado */
+.filter-select:focus {
+    outline: none; /* Elimina el borde de enfoque predeterminado */
+    box-shadow: 0 0 8px #ccc; /* Sombra azul */
+}
+
+
+/* Estilo para las opciones cuando se pasan por encima (hover) */
+.filter-select option:hover {
+    background-color: #ccc; /* Fondo gris claro */
+    color: #333; /* Texto gris oscuro */
+}
+
+.styled-label {
+    font-size: 16px; /* Tamaño de fuente */
+    font-weight: bold; /* Negrita */
+    color: #333; /* Color del texto */
+    margin-right: 10px; /* Espacio a la derecha */
+    padding: 5px 10px; /* Espaciado interno */
+    background-color: #f0f0f0; /* Color de fondo */
+    border-radius: 5px; /* Bordes redondeados */
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1); /* Sombra */
+    display: inline-block; /* Asegura que el label se comporte como un bloque en línea */
+}
+
